firmware: ch32x: rename header
